requirements.txt
python-telegram-bot==13.15
python-dotenv==1.0.0
redis==4.5.5
vk-api==11.9.9
